Fix for windows, GetGrowthStats was (silently) failing because sh was being used to execute the command.
For now force the use of  bash in windows (and only in windows). can't just call git for now, since since the executed command uses a pipe, so we will go with bash for now.

This shouldn't be an impediment if git is installed so should be bash.  We might later remove this dependency from bash